refactor: isolate an inner Consensus object
