+++ network/src/consensus/consensus.rs
@@ -35,14 +35,8 @@ use std::{
 pub struct Consensus<S: Storage> {
     /// The node this consensus is bound to.
     node: Node<S>,
-    /// The storage system of this node.
-    storage: Arc<MerkleTreeLedger<S>>,
-    /// The memory pool of this node.
-    memory_pool: Arc<Mutex<MemoryPool<Tx>>>,
-    /// The consensus parameters for the associated network ID.
-    consensus_parameters: Arc<ConsensusParameters>,
-    /// The DPC parameters for the associated network ID.
-    dpc_parameters: Arc<PublicParameters<Components>>,
+    /// The core consensus objects.
+    consensus: snarkos_consensus::Consensus<S>,
     /// If `true`, initializes a mining task on this node.
     is_miner: bool,
     /// The interval between each block sync.
@@ -59,20 +53,14 @@ impl<S: Storage> Consensus<S> {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         node: Node<S>,
-        storage: Arc<MerkleTreeLedger<S>>,
-        memory_pool: Arc<Mutex<MemoryPool<Tx>>>,
-        consensus_parameters: Arc<ConsensusParameters>,
-        dpc_parameters: Arc<PublicParameters<Components>>,
+        consensus: snarkos_consensus::Consensus<S>,
         is_miner: bool,
         block_sync_interval: Duration,
         transaction_sync_interval: Duration,
     ) -> Self {
         Self {
             node,
-            storage,
-            memory_pool,
-            consensus_parameters,
-            dpc_parameters,
+            consensus,
             is_miner,
             block_sync_interval,
             last_block_sync: RwLock::new(Instant::now()),
@@ -89,25 +77,25 @@ impl<S: Storage> Consensus<S> {
     /// Returns a reference to the storage system of this node.
     #[inline]
     pub fn storage(&self) -> &Arc<MerkleTreeLedger<S>> {
-        &self.storage
+        &self.consensus.ledger
     }
 
     /// Returns a reference to the memory pool of this node.
     #[inline]
     pub fn memory_pool(&self) -> &Arc<Mutex<MemoryPool<Tx>>> {
-        &self.memory_pool
+        &self.consensus.memory_pool
     }
 
     /// Returns a reference to the consensus parameters of this node.
     #[inline]
     pub fn consensus_parameters(&self) -> &Arc<ConsensusParameters> {
-        &self.consensus_parameters
+        &self.consensus.parameters
     }
 
     /// Returns a reference to the DPC parameters of this node.
     #[inline]
     pub fn dpc_parameters(&self) -> &Arc<PublicParameters<Components>> {
-        &self.dpc_parameters
+        &self.consensus.public_parameters
     }
 
     /// Returns `true` if this node is a mining node. Otherwise, returns `false`.
@@ -129,7 +117,7 @@ impl<S: Storage> Consensus<S> {
     /// Returns the current block height of the ledger from storage.
     #[inline]
     pub fn current_block_height(&self) -> u32 {
-        self.storage.get_current_block_height()
+        self.consensus.ledger.get_current_block_height()
     }
 
     /// Checks whether enough time has elapsed for the node to attempt another block sync.
@@ -149,6 +137,6 @@ impl<S: Storage> Consensus<S> {
     }
 
     pub fn max_block_size(&self) -> usize {
-        self.consensus_parameters.max_block_size
+        self.consensus.parameters.max_block_size
     }
 }
